chore: rename confirm cards
